Ignore unknown tags when generating profiles; disconnect main profile from ftp-master-auto-reject.
During the release process, profiles are regenerated before shipping,
but the information about auto-reject tags obtained from FTP Master
can be out of date. This ignores unknown tags in that list when the
profiles are generated.

When new main profiles are generated, they will also no longer extend
the ftp-master-auto-reject profile. It is not clear why that was so in
the first place. The main profile now stands on its own.

This commit changes the script that generates the profiles. It does
not actually modify any profiles.