Reverse exit statuses for program error and policy violations. (See: #709932)
The previous use of exit status 1 for policy violations caused a
series of problems. For the most part it was difficult to tell policy
violations from errors in programs like 'sudo'.

This reverses the exit codes. In the future, program errors will
return status 1. Policy violations will return status 2.  Users will
also have more control over the latter with the upcoming '--fail-on'
command-line option.

An additional benefit is that it enables the use of 'die' throughout
Lintian. It makes it unnecessay to trap exceptions from modules that
do use 'die'. Fewer Perl evals, in turn, make debugging easier.