Fix several 'Todo' tests using literal match strategy. (Closes: #921112)
These tests were supposed to be fixed when Match-Strategy: literal
became available. The 'tainted' warning from Lintian can be compared
literally. No further calibration of the output is needed.

Due to the Match-Strategy: literal facility, the test suite can handle
program warnings that are not tags. The related bug is closed.