In CI pipeline, pin installation of debhelper to -backports; should restore pipeline.
The stable-bpo pipeline has been failing because the virtual package
debhelper-compat (= 13) cannot be installed. It may be a limitation of
apt. The version is available in buster-backports.

This commit uses the target release mechanism to install debhelper
from stable-backports before any build prerequisites are resolved. It
reads the release code name from /etc/os-release to avoid adjustments
in the future.

The advance installation of debhelper is allowed to fail. That should
cover the brief period after a new release when the backports repo is
not yet available.

The apt behavior at issue here probably deserves a bug report, but the
package has over eight hundred open bugs. Some research would be
needed to avoid a duplicate filing. The maintainer also has not been
very friendly. Bug#941347 may be related.

Exports the environment variable DEBIAN_FRONTEND=noninteractive so
that it does not need to appear in individual statements.

Enables tracing of shell commands as they are executed.

This restores the stable-bpo pipeline in a private Gitlab repo.