Use Email::Address::XS to parse fields for Maintainer, Uploaders and Changed-By.
Replaces some basic checks in Lintian::Maintainer with a UTF-8 capable
stock module. Renames some tags for a better indication to users.

Introduces classification tags for the maintainer and uploaders fields
that are useful for the new web reporting framework.

May enable Bug#485705 to be solved properly, but the point is probably
moot.