Accept some environmental variables as valid keys in the configuration file. (Closes: #961800)
The new parser for configuration files allows some environment
variables to be set but they were rejected as invalid keys. This
accepts the variables as valid keys, as well.

As a bonus, this commit adds a debug message when the environment is
modified via the configuration file.

Thank you to gregor herrmann for the report, and for the suggested
fix!