Clarify tag names related to shlibs control and shared libraries. (Closes: #534938)
Many tags for shared libraries use 'shlib' in their names but that was
too brief and also to similar to the Debian shlibs control file. Even
for experienced operators, the meaning was often unclear.

Some tags have many overrides, more than those in commit bab2ca64
(also see Bug#922544). The tag rename facility should that make the
process a little easier for maintainers. The old names were recorded
in the new tag declarations as previous names.

The high number of overrides also raises other questions. Do the tags
function as intended? Do the odd names they had until now obscure the
problem to maintainers lost patience and overrode?

The tests will be adjusted in a nearby, future commit. They were
searched for the old names, but some may fail on architectures other
than amd64.

This renames the following 24 tags as indicated. The numbers in
parentheses are the number of overrides in a recent archive-wide run
and the total number of tags:

( 446 of  531) binary-or-shlib-defines-rpath -> custom-library-search-path
( 106 of  210) dev-pkg-without-shlib-symlink -> lacks-unversioned-link-to-shared-library
(   0 of    0) duplicate-entry-in-shlibs-control-file -> duplicate-in-shlibs
(   2 of    3) ldconfig-symlink-missing-for-shlib -> lacks-versioned-link-to-shared-library
( 142 of  405) non-dev-pkg-with-shlib-symlink -> link-to-shared-library-in-wrong-package
( 135 of  138) no-shlibs-control-file -> no-shlibs
(   0 of    0) pkg-has-shlibs-control-file-but-no-actual-shared-libs -> empty-shlibs
(   0 of    0) pkg-has-symbols-control-file-but-no-shared-libs -> empty-shared-library-symbols
(1327 of 3413) shared-lib-without-dependency-information -> shared-library-lacks-prerequisites
( 180 of 3225) shlib-calls-exit -> exit-in-shared-library
(   3 of    5) shlib-in-multi-arch-foreign-package -> shared-library-is-multi-arch-foreign
(  25 of   25) shlib-missing-in-control-file -> ships-undeclared-shared-library
(  28 of   29) shlib-missing-in-symbols-control-file -> shared-library-symbols-not-tracked
( 783 of  792) shlibs-declares-dependency-on-other-package -> distant-prerequisite-in-shlibs
(   0 of    0) shlibs-uses-obsolete-relation -> outdated-relation-in-shlibs
(   1 of    1) shlib-with-bad-permissions -> odd-permissions-on-shared-library
(  89 of   90) shlib-with-executable-bit -> shared-library-is-executable
(  15 of 1245) shlib-with-executable-stack -> executable-stack-in-shared-library
(   0 of    2) shlib-with-non-pic-code -> specific-address-in-shared-library
(   0 of   62) shlib-without-PT_GNU_STACK-section -> shared-library-lacks-stack-section
( 126 of  170) shlib-without-versioned-soname -> shared-library-lacks-version
(   0 of    0) symbols-declared-but-not-shlib -> symbols-for-undeclared-shared-library
(   4 of    6) unused-shlib-entry-in-control-file -> shared-library-not-shipped
(   0 of    0) unused-shlib-entry-in-symbols-control-file -> surplus-shared-library-symbols