Use Email::Address::XS to parse all address components. (Closes: #961924)
The bug occurred because the maintainer field was parsed with multiple
methods. This commit uses Email::Address::XS throughout.

The ->format method from Email::Address::XS surrounds the phrase with
double quotation marks, while the parser in Lintian::Maintainer could
not handle them.

A false-positive test case that mimicks bwbasic from the bug report
was added in the previous commit. The erronous tags are corrected in
this commit.

Also adjusts the remaining tests for the new output, which now
emphasizes address components with issues. The full field values were
reported previously.