Replace configuration option LINTIAN_PROFILE with 'profile'. (Closes: #961855)
This bug is about an error in the processing of configuration files.
The LINTIAN_PROFILE value was added to the environment but not to the
program options. The profile was therefore never read.

After some reflection, the current LINTIAN_PROFILE option was replaced
with the more intuitive 'profile'. Like most other options it now
mirrors those given on the command line. The value is no longer added
to the environment.

As a side note, the bug report was not reproducible in all details.
Perhaps the profile name should have been prefixed with 'debian', as
in:

    LINTIAN_PROFILE = debian/dmo

Either way, the line should read:

    profile = debian/dmo

The old option works again, but now generates a warning.

Thanks to Christian Marillat for reporting the bug! Sorry about the
inconvenience.