Combine tags for missing native and non-native changelogs in installation packages. (Closes: #935072)
The two tags were combined into a handy tag called no-changelog.  The
context shows the name under which such a file should be installed.

Neither of the tags appeared in a recent archive-wide run. The old
names were listed in the new tag's rename facility.

The expected tags in the test suite were also adjusted herein.