Split check changes-file into the appropriate components in fields.
The check was orthogonal to the design of all other checks. Since the
automatic facet 'changes' became available for when checks are invoked
on *.changes files, there was no advantage to a check for such files.

This splits the check into new but smaller checks in fields or adds
to the functionality of checks that exist there already.

Gbp-Dch: ignore