Parse lists of distributions in changes files.
The value for the Distribution field can be a list, although Debian's
archive software can only handle a single value. Policy 5.6.14 states:

    "this [field] contains the (space-separated) name(s) of the
    distribution(s) where this version of the package should be
    installed. ... The Debian archive software only supports listing
    a single value."

This commit parses the distribution fields as lists of values.

Some of the expected tags in tests changed due to small modifications
in the output format for the context.