Use UTF-8 codepoint for copyright sign in copyright files. (Closes: #961975)
After some reflection, the UTF-8 codepoint for the copyright sign was
use instead of the suggested \N{COPYRIGHT SIGN}.

Being able to do so is in fact the sole reason for 'use utf8;' in all
of Lintian's source files. We use the same codepoint in many of our
own copyright notices.

For gnome-shell-extension-bluetooth-quick-connect_13-1_all.deb,
Lintian no longer produces any tags.

Thanks to Simon McVittie for this well-researched and helpful bug
report!

As an aside, the regex may have other issues. For example, the use of
the character class around (C) makes the expression overly broad.

Also, the fields should be parsed separately (instead of the entire
contents). This check is scheduled for a rewrite.