Remove command line option --fail-on from the settings in configuration files. (Closes: #962157)
The configuration file supports a limited number of settings for end
users. They all relate to the display (and one allows selecting the
number of jobs).

The --fail-on command option, on the other hand, is used only in
automated settings.  No one on the team that implemented it plans to
use it in a configuration file. It was a mistake to include that
option among those available in configuration files.

For automated settings, the --fail-on command-line option can be
conveniently enabled in a transparent and straightforward manner. It
should be added by the invocant as '--fail-on error'.

Users who rely on the exit status are furthermore advised to examine
their scripts. The exit codes for program errors and the conditions
selected via --fail-on are now reversed.

The new option was introduced at the same time to miminize the overall
inconvenience. Thank you for using Lintian!