Flag the use of acute accents in man pages. (Closes: #554897)
A scan of man pages showed that the offending character sequence is
also used when requesting string variables with the 'ds' macro:

/usr/share/man/man8/update-default-wordlist.8.gz:.    ds ' \\k:\h'-(\\n(.wu*8/10-\*(#H)'\'\h"|\\n:u"
/usr/share/man/man8/gdm3.8.gz:.    ds ' \\k:\h'-(\\n(.wu*8/10-\*(#H)'\'\h"|\\n:u"
/usr/share/man/man8/remove-default-ispell.8.gz:.    ds ' \\k:\h'-(\\n(.wu*8/10-\*(#H)'\'\h"|\\n:u"
/usr/share/man/man8/pg_updatedicts.8.gz:.    ds ' \\k:\h'-(\\n(.wu*8/10-\*(#H)'\'\h"|\\n:u"
/usr/share/man/man8/apparmor_parser.8.gz:.    ds ' \\k:\h'-(\\n(.wu*8/10-\*(#H)'\'\h"|\\n:u"

That use was allowed, but the check otherwise assumes that man pages
never use acute accents. That is probably too strict. At the same
time, the erroneous use of the sequence is common in English language
man pages.

We hope there is a net benefit to this tag from the outset. Please
report a bug if the check does not work for your man page.