Drop Email::Valid as obsolete; use regex to check emaisl for fully qualified hosts. (See: #962277)
In an unusual uploader address, Email::Valid choked on the UTF-8
sequence RIGHT-TO-LEFT OVERRIDE. The team consensus is that such an
address should be valid.

Email::Address::XS already allowed it, but Email::Valid did not.

The package was Ansgar's 'colorful', which incorporated terminal
escape sequences, including blinking. It was converted to two test
cases in a recent commit.

The Email::Valid module implements the older RFC 822. It was recenly
used only to check that host names are fully qualified.  Replaces that
last remaining use with a simple regular expression.

Also drops libemail-valid-perl from d/control.