Flag ANSI escape characters in all fields of primary packaging control files. (See: #962277)
This check was requested in response to Ansgar's 'colorful' test
package. Please check out the test case included here or look at
commit ad4c446e for the interesting and lighthearted creation.

Thank you to Paul Wise for the suggestion.

Ansgar's package contains ANSI escape sequences for many colors. The
Maintainer field from that package serves as the test case here.  It
was moved from the check fields/maintainer, where it served briefly as
a false positive.

This check applies to all fields in d/control in debs and udebs, in
*.dsc source files, in *.changes files and in *.buildinfo files.

Ships regenerated profiles that include the new check.