Issue 'required-field' and 'recommended-field' per policy 5 instead of field-specific tags.
Implements all required and recommended field specifications in Policy
section 5. The field names from policy are furthermore kept in one
place for easier maintenance.

As one caveat, policy section 5.2 states that the fields Section and
Priority are recommended in the source paragraph as well as in the
binary paragraphs, but in the author's opinion it does not make sense
to flag them in the latter.

The fields in the source paragraph provide the default for the binary
package paragraph. Moreover, having the fields in both places would
trigger the tag binary-control-field-duplicates-source, which would be
confusing.

Combines the following tags into 'required-field', and issues
additional ones per policy when fields are missing:

  - no-maintainer-field
  - no-architecture-field
  - no-version-field
  - no-standards-version-field
  - no-package-name
  - package-has-no-description
  - no-description-in-changes-file

Combines the following tags into 'recommended-field', and issues
additional ones per policy when fields are missing:

  - no-section-field
  - no-priority-field
  - no-section-field-for-source
  - no-urgency-in-changes-file