In check menus, title-case the hash keys representing field labels to make the tag context more human-readable.
This is part of a larger commit series that will keep all hash keys
for Debian RFC822 style config files in title-case. It makes for
better tag and error reporting, and avoids the cumbersome task to
convert labels whenever they are shown.

As an added bonus, the code looks better, too.