Add a classification tag for field contents in primary package control files.
The new tag 'trimmed-field' provides access to major classification
data. It is expected to be useful for the lintian.d.o web site, and
perhaps Debian Janitor.

The data provided is the trimmed field contents. Some output formats
replace literal newlines with \n but the information should otherwise
be true to the original, except that leading whitespaces are trimmed.