Rename tag for mailing list on Alioth; convert to classification tag. (Closes: #962448)
There was concern among Lintian's maintainers to exempt individual
teams. No other teams had been exempted to date, even though multiple
teams will continue to use the old address.

In closing this bug, the Java team was not exempted from the tag as
requested. Instead, the tag was converted to a classification tag,
which means it states a fact. The tag was also renamed to reflect the
facts more accurately. We hope that this solution is acceptable to the
Java team.

The tag is no longer experimental.

On the renaming, 'obsolete' was too strong a word to describe an
address that is discouraged but will be serviced going forward.
Alioth was also the only system that triggered the tag. The new name
reflects both.