Move all tests from t/tags to t/recipes.
Corrects an old misnomer (remember the old suite 'tags'?) that has
survived until now. We refer to these files as test specifications,
but the latter was simply too long.

If you encounter problems rebasing old branches, please experiment
with git's rename limits:

    git config merge.renameLimit 999999
or
    git config diff.renamelimit 999999

This change is sure to create pain for someone. At the same time, it
had been delayed for many years, and was well overdue.

The test executables will be adjusted in the following commit.