Do not allow unsafe symlinks in autopkgtests; limit and rename tag. (Closes: #954860)
The bug report was correct. The tag source-contains-unsafe-symlink did
in fact duplicate part of absolute-symbolic-link-target-in-source.  A
maintainer's initial objection to the filing was mistaken.

While the cited test case performed as described, the argument did not
consider that source-contains-unsafe-symlink exempted autopkgtests
from unsafe symlinks. The cited test case provided exactly those
links.

As a result of ths commit, autopkgtests are no longer allowed to use
unsafe symlinks in source packages.

Prior maintainers also had doubts about allowing such links in
autopkgtests. This comment was in a test that provided such links:

>   # Lintian ignores these - not sure to what extend it makes sense
    ln -s "../../$LINK_TARGET" "$SOURCE_ROOT/tests/relative-escape-but-ignored"
    ln -s "/../../$LINK_TARGET" "$SOURCE_ROOT/tests/absolute-escape-but-ignored"

Autopkgtests sometimes use links to /dev/null to designate no-ops.
They are no longer allowed in source packages. Please modify the
relevant documentation if you know where the practice originated.

This commit effectively splits source-contains-unsafe-symlink into
two. Absolute links trigger absolute-symbolic-link-target-in-source, a
warning, while relative links that cannot be resolved now trigger the
tag wayward-symbolic-link-target-in-source, an error.

The tag wayward-symbolic-link-target-in-source is no longer issued for
absolute link targets. That removes the overlap that gave rise to the
bug report.

Thanks to Raphaël Hertzog for making this report!