Broaden tag descriptions for lower custom compression settings; rename tags. (Closes: #909696)
Based on a suggestion near the end of the bug report, this commit
expands the tag descriptions to explain why someone would use settings
for lower compression settings.

The reporting party wished to continue to support Wheezy, where xz was
not available. (Support for xz appeared gradually; the Dpkg changelog
did not state clearly when the algorithm was supported everywhere.)

The bug report is now nearly two years old. Perhaps the problem
drifted into history. Otherwise, an override may be appropriate.

A more thorough fix for the bug would require better parsing.

One tag was renamed to remove the 'should-not-use' formula, as in
other renaming efforts. The second tag was renamed to be similar.