Reset process umask before unpacking source tarballs. (Closes: #945869)
The previous commit fully exposed a Dpkg bug (#796257) that causes
Lintian results to differ depending on the user's umask.

The tests in the previous commit were calibrated using a umask of 000
when sources were unpacked using dpkg-source. The tests fail if the
invoking user sets any other umask.

This commit always sets the umask to 000 before calling dpkg-source.
It finally makes Lintian results consistent regardless of the user's
umask.