Emit a classification tags with octal file permissions for all files sources and installables. (See: #945869, #796257)
This new tag and the associated tests expose a Dpkg bug (#796257) that
causes Lintian results to differ depending on the user's umask.

The tests were calibrated using a umask of 000 when the sources are
unpacked using dpkg-source. The Salsa CI pipelines should fail if the
runners impose any other kind of umask, such as 022.

The commit following this one sets the umask to 000 before calling
dpkg-source. That should restore the CI pipelines and finally make
Lintian results consistent regardless of the user's umask.