Use case-sensitive field names in control files, package relations, debconf templates and autopkgtest files.
From now on, Lintian will look at literal fields names. Policy 5.1
permits arbitrary capitalization in many of these fields, but that
complicates the visual presentation and the examination of same.

A future commit may implement case-invariant access to the field
content while keeping the original names around. This is planned
sooner rather than later, because it will affect some users.

After the switchover, Lintian will no longer work properly with fields
that do not employ the familiar title-case.

While counter-intuitive, there is one user-friendly aspect of this
change: It removes the title-casing used in many checks to make the
tag messages more attractive.  Now the original capitalization is
shown, which is presumably preferred by the user.