Fix debconf tests that use the Deb822Parser to generate maintainer scripts.
One has to regenerate the test packages for this build error to show.
The author did not do so prior to the last commit.

Using the full parser is unnecessary when looking for single lines.
The tests will be rewritten in the near future to prevent parser
changes from causing this type of problem again.

This should restore the Salsa CI piplines.

Gbp-Dch: ignore