Do not use bash when syntax checking maintainer scripts using /bin/sh. (Closes: #954459)
The 'scripts' check will not find bashisms when /bin/bash provides
/bin/sh.

It is so, because the routine check_script_syntax is called with the
interpreter named in the shebang. When the script uses /bin/sh and the
latter is provided by /bin/bash, the syntax check runs in bash, which
will not find bashims.

Another shell should be used for the syntax check for /bin/sh. This
commit uses dash for maintainer-shell-script-fails-syntax-check.

Thanks to Paul Wise for helping and explaining the bug report! Paul
also had a vague recollection that 'sh -n' was unsafe. It has yet to
be explored.

Thanks to Olly Betts for confirming that dash is Essential: yes.