Point out variant spellings of field names in control files.
Introduces a tag with suggestions for more common spellings.  Lintian
recently became case aware. Due to the potential for controversy, the
tag is at the pedantic level.

The primary objective of this check was to make sure we do not lose
the nifty title-case algorithm that is no longer needed.

Provides three test cases. Regenerates profiles.