Use case-sensitive Deb822 parser for profiles, tags and checks.
The deb822 format allows for flexible capitalization, but the parser
was recently modified to capture the literal field names. It was done
for presentation purposes. We retired some complex and repetitive
output formatting.

Case invariance will return for the parsing of package control files.
It is unclear if it is also needed for the parsing of profiles, tags
and checks.