Use case-sensitive Deb822 parser in internal harness tests.
The deb822 format allows for flexible capitalization, but the parser
was recently modified to capture the literal field names. It was done
for presentation purposes. We retired some complex and repetitive
output formatting.

Case invariance will return for the parsing of package control files.
It probably won't return to the internal harness tests. They are under
our sole control and, as tests, deserve to be picky.