Use case-sensitive parsing in private scripts.
The deb822 format allows for flexible capitalization, but the parser
was recently modified to capture the literal field names. It was done
for presentation purposes. We retired some complex and repetitive
output formatting.

Case invariance will return for the parsing of package control files.
It is unclear if it is also needed for these private scripts.

Some of these scripts, such as tag-stats, appear to have fallen out of
use.