Use case-sensitive Deb822 parser in for readelf data in Processable::Objdump.
The deb822 format allows for flexible capitalization, but the parser
was recently modified to capture the literal field names. It was done
for presentation purposes. We retired some complex and repetitive
output formatting. The readelf parser happens to use the same code.

Case invariance will return for the parsing of package control files.
It is unclear if it is needed for readelf output.