Add new tag breakout-link for such links in /usr/lib. (Closes: #243158)
This check currently requires that links in /usr/lib always point to
targets in the same directory or a subdirectory thereof.

Lateral connections to the same level in /usr/lib were not allowed
based on the typical multi-lib layout in which that might point to
another architecture.

Provides a test case. Regenerates profiles.