Rewrite method to find missing sources; performs like previous code. (See: #962583)
A rephrasing was needed to bring clarity into the decision-making as
to which sources are considered missing. This code performs like one
before, and has the same shortfalls.

For example, all files in debian/missing-sources are exempt from the
check for source-is-missing. (I believe it was unintentional, as
explained below.)  Those files should, in fact, be subject to greater
scrutiny. It will also annoy users who had hoped to escape the check
by 'declaring' that the files are sources.

The following lines are responsible. This ones exempts all entries
using the full path below the missing-sources folder:

    my $fullpath
      = $index->resolve_path('debian/missing-sources/' . $file->name);
    return $fullpath
      if defined $fullpath;

While this one does so for entries using just the basename:

    my $abbreviated
      = $index->resolve_path('debian/missing-sources/' . $file->basename);
    return $abbreviated
      if defined $abbreviated;

A better way would be to the two files to the list of candidates.  It
will require that they do not point to the file being inquired about.
That lifts the accidential exemption.

The connection is not obvious, so the exemption was probably
unintentional.  It will be fixed soon. Files in debian/missing-source
will start to be subject to the source-is-missing tag.

Gbp-Dch: ignore