Fix bug for consecutive slashes in File::Path::resolve_path.
The routine resolve_path traverses path components recursively. Until
now, it did not work properly with paths containing consecutive
slashes.

The rewrite of the indices required, for internal consistency, that
absolute paths are only used with anchored indices. The method
resolve_path aborts otherwise.  Unfortunately, it also meant the
routine underperformed when used recursively.

The next path element is calculated with an regular expression. That
expression took a additional slash to avoid the problem, but should
have taken all of them!  Otherwise, the next recursive call on
unanchored indices aborts as unresolvable. On anchored indices, a new
absolute reference is calculated, which is also incorrect.

This commit modifies the regular expression to remove all slashes
following the current path element before calling the method again
recursively.

Double slashes occur primarlily due to concatenation. Much of that is
manual, as in 'my $path = "$directory/$name";' when $directory ends in
a slash, but it was still a serious source of errors. Now it is fixed.

This bug appeared when working on the check cruft to address
Bug#962583.

Gbp-Dch: ignore