Stop exempting files in debian/missing-sources from source-is-missing. (See: #962583)
Under the previous code, files in debian/missing-sources were exempt
from the check for source-is-missing. The reason was not obvious, and
it was most likely unintentional. This removes the exemption.

From now on, files in debian/missing-sources will be tested for
source-is-missing.

There could be a philosophical debate as to whether the files should
be checked at all. After all, the package maintainer declared that
they are valid and human-readable.

At the same time, files with such a declared status should probably
withstand even greater scrutiny. Surely, maintainers who had hoped to
escape the check by 'declaring' that files are sources will disagree
with this change.

Here is why the files were not checked: Unlike all other candidates,
the two locations at issue here did not have to point to files other
than the one being inquired about to produce a positive hit. But
because debian/missing-source/$basename equals the full path for any
file in that folder, a source was thought to exist when it was simply
the same file.

It was probably a programming oversight. It is fixed here.

There was also a brief explanation in the previous commmit.