In Makefile for source tests, copy original sources for native.
This change did not expose any issues in the test suite. It is
probably because so few native source packages are generated, and even
fewer contain orig content. Some generate their orig content with a
pre-build script.

The upload-native template, which is used more frequently, also
produces source packages, but it uses another Makefile.

Gbp-Dch: ignore