Issue tag for repeated path segments in installed files. (Closes: #950052)
The tag shows path segments that repeat in locations of installed
files. It prevents, for example, the accidental shipment of a file in
/usr/lib/lib.  All shipped paths are examined, including links.

Path segments do not have to be adjacent. Just using a segment again
in a path will trigger the tag. Multiple duplicates in the same path
appear separately.

Basenames are disregarded. That prevent false positives for files
named after their encapsulating directories, possibly a valid use.

Provides a test case. Regenerates profiles.