Remove old methods from Lintian::Processable, Lintian::Group, and Lintian::Pool.
The clear_cache method was kept around to help free working memory,
but became a no-op and fell out of use.

The info method went through an even greater number of iterations and
ultimately pointed to the processable itself because classes were
combined.

The type method in Lintian::Group returned a fixed sting, but there
are no subclasses. It has not been used in years.

Removes the methods and their invocations and counterparts elsewhere.

Gbp-Dch: ignore