Restore case-invariant access to control fields, per policy 5.1.
Commit 4cdf7d21 began to require that fields are accessed with the
exact capitalization used in control files. For packages using variant
spellings, it meant checks would not run properly because they could
not find fields.

Restores the case-insensitive access to field contents.

Raw fields are stored according to their original capitalization.
During a lookup, names are converted to lower case.  The legend
provides the original capitalization. That is then used to retrieve
the field contents.

For unfolded fields, hash keys were converted to lower case. No legend
is needed. Relations are also stored in lowercase. Like the other
fields, they can be accessed using any capitalization.

The regression was a consequence of switching to a case-sensitive
parser for deb822 files. That change was made to improve tag output.

Corrects an expected tag accepted in error in commit 4cdf7d21.