Add internal harness test requiring that test recipes are calibrated. (See: #962671)
Requires that an 'eval/tags' or an 'eval/literal' file is present in
test recipies. The files are indicators that a test was calibrated.

Without a calibration file, the test suite assumes that no tags (or no
output) are expected, but that also makes it harder to analyze test
recipes for their diagnostic value.

Gbp-Dch: ignore