Add an internal harness test requiring a declared diagnostic value for tag-related test. (Closes: #962671)
From now on, tag-related test recipes must look for expected tags or
declare a Test-Against clause for false positives.

Due to the high number of incomplete tests, an exception list was
added similar to the one for untested tags. It will take some time to
determine the diagnostic purpose for each test on the list.