Expand internal test script to validate the HTML generated from tag descriptions.
Uses the well-known tidy program to look for warnings or errors in the HTML
generated from the Markdown in tag descriptions.

Checking the HTML is superior to checking the Markdown because Markdown can
contain HTML tags and entities. It is the result that matters.

Another advantage is that the description contains additional information,
such as a statement about references, if any are available. That extra
information is checked here, as well.

Gbp-Dch: ignore