Also loop over patched source files and control files in Lintian::Check.
Until now, the check facilities only looped over installed files. This
adds new entry points for patched source and control files:

   setup_X
   files_X
   breakdown_X

where X is the name of the index, as in 'patched' or 'control'.

To avoid name confusion, new entry points are provided for installed
files, but the old ones continue to work:

   setup_installed
   files_installed
   breakdown_installed

Gbp-Dch: ignore