Move checks from lintian to debian/lintian-overrides. (See: #950052)
In installation packages, checks ship in /usr/share/lintian. For
checks concerning overrides, the path segment 'lintian' appeared
twice. It triggered the new tag repeated-path-segment.

This commit moves the affected two checks to debian/lintian-overrides.
It implements option 2 from the bug.

This does not move the tests related to the internal 'lintian' check
(soon to be obsolete). The test are not shipped, however, and there is
no issue with their installation path, although that could change when
Bug#926409 is addressed.

Regenerates profiles.