Rename new item-looping entry points in checks; not all items are files.
Changes the name of the recently introduced entry point 'files_X'
(where X is the name of the index) to 'visit_X'.

The new name reflects the looping nature and accounts for the fact
that not all items visited are files. Among them are many folders and
links.

The old entry point 'files' retains the name for backward
compatibility. It is used in some team checks.