Provide script diagnostics in control index.
The control index uses a custom mechanism for packaging scripts, but
the scripts mechanism provides much of the same. Enables the scripts
mechanism for the control index.

The goal is to reduce the duplication between both mechanisms. (As
primary differences, control scripts can be ELF executables and are
filtered by name.) Moves Lintian along part of the way.

The functionality is used by the check files/encoding.