Provide a test case with national encoding in packaging scripts.
The functionality in this test was not previously tested.

Another test, control-in-western-encoding, ships a control file in
western encoding but Lintian refuses to add the resulting installation
package (type 'all') to the pool.

A fallback parser is needed for packaging control files that are not
encoded in UTF-8. It is in the works. When complete, an additional tag
will appear in the other test.

The tags in this test will be unaffected.

Gbp-Dch: ignore