Drop parse_dpkg_control from Lintian::Deb822Parser.
The routine was only used in two places, and suboptimally so.

Both consumers were better served with similar routines. That dropped
the last remaining uses of parse_dpkg_control. The routine and its
lowercase cousin are gone and no longer provided.

The utility program ./private/refresh-archs used to generate many
warnings, but they no longer make it to the terminal. The warnings
provided confidence to the user that the program, which takes about a
minute, was working on something.

 Now a brief note is printed instead:

    Working on architecture netbsd-arm.

The warnings looked like this:

    dpkg-architecture: warning: specified GNU system type arm-netbsd does not match CC system type x86_64-linux-gnu, try setting a correct CC environment variable

Gbp-Dch: ignore