Parse package control files even if they have an odd encoding.
Lintian currently disregards packages whose primary control files are
not encoded in UTF-8. This commit causes the files to be parsed,
however imperfectly, even when they are not valid UTF-8

Relaxing the parser makes it possible to provide tags for those
packages. That includes the important warning about the control file
being in a national encoding!

Produces additional tags in two tests. The target test was among them.