In two tests, replace files ignored by dpkg-source; their absence caused FTBFS.
Restructures two tests so the Lintian source does not ship files that
are on the ignore list for dpkg-source. The upload for 2.81.0 suffered
an FTBFS on the buildd network because two files were absent.

Both files were dummy shared objects named with *.so extensions. In
one case, the extension was changed; in the other, the file is now
being created dynamically by d/rules.

Here is the full list from 'dpkg-source --help':

  -I, --tar-ignore[=<pattern>]
                           filter out files when building tarballs
                             (defaults to: -I*.a -I*.la -I*.o -I*.so
                          -I.*.sw? -I*/*~ -I,,* -I.[#~]* -I.arch-ids
                          -I.arch-inventory -I.be -I.bzr -I.bzr.backup
                          -I.bzr.tags -I.bzrignore -I.cvsignore -I.deps
                          -I.git -I.gitattributes -I.gitignore
                          -I.gitmodules -I.gitreview -I.hg -I.hgignore
                          -I.hgsigs -I.hgtags -I.mailmap -I.mtn-ignore
                          -I.shelf -I.svn -ICVS -IDEADJOE -IRCS -I_MTN
                          -I_darcs -I{arch}).