Drop test no longer building due to new restraints on version strings. (See: #953554)
For two days, the Gitlab CI for unstable has been failing.  Lintian
was unable to build a test package. The error message was:

    dpkg-source: error: non-native package version does not contain a revision

From the Dpkg changelog, it seem that its maintainers imposed new
restrictions on version strings. The entry from 27 Jun 2020 states:

dpkg (1.20.1) unstable; urgency=medium

  [ Guillem Jover ]
    ...
    - Dpkg::Source::Package::V1: Check version format matching source format.

The changes can be examined in greater detail in this commit message:

    https://salsa.debian.org/dpkg-team/dpkg/-/commit/aa605ebd31e7ecc212ade755068cde74ebde2cb3

This commit removes the offending test, which should restore the
Gitlab pipeline.